refactor(server): move workspace files into workspace/ subfolder
The server package has two domains — sync and workspace — but only sync
had its own subfolder. This moves the workspace files into workspace/ to
mirror the sync/ organization and make the two domains symmetrical.
