+++ packages/server/src/index.ts
@@ -1,2 +1,2 @@
 export { createServer, DEFAULT_PORT, type ServerOptions } from './server';
-export { createWorkspacePlugin } from './workspace-plugin';
+export { createWorkspacePlugin } from './workspace';
