+++ packages/server/src/workspace/index.ts
@@ -1,0 +1,2 @@
+export { collectActionPaths } from './actions';
+export { createWorkspacePlugin } from './plugin';
